Lift state up to AppEntry and Pass DrugData as context
