+++ src/HomeComponents/StockComponents/Stock.jsx
@@ -1,3 +1,11 @@
+import { useContext } from "react";
+import { DrugData } from "../../DrugDataContext";
+
 export default function Stock() {
-	return <div>Stock</div>;
+	const { drugs, input } = useContext(DrugData);
+	const [searchInput] = input;
+	const [drugData] = drugs;
+
+	console.log(drugData);
+	return <div></div>;
 }
